test: Moved specific backbone unittests
